backend/workers/compatibility/requirements.txt
azure-servicebus
requests
PyYAML